ac {gclus}	R Documentation

Clustering coefficients from package cluster.

Description

Computes clustering coefficients from cluster, where x and y give the object coordinates.

Usage

ac(x, y, ...)
sil(x, y, groups, ...)

Arguments

x
is a numeric vector.

y
is a numeric vector.

groups
is a vector of group memberships, used by sil only.

...
are passed to agnes in ac and to dist in sil.

Details

ac - Computes clustering coefficient from agnes{cluster}.

sil - Computes the silhouette coefficient from from package cluster.

Value

The clustering coefficient is returned.

Author(s)

Catherine B. Hurley

References

Kaufman, L. and Rousseeuw, P.J. (1990). Finding Groups in Data: An Introduction to Cluster Analysis . Wiley, New York.

See Also

agnes, silhouette, dist.

Examples

x <- runif(20)
y <- runif(20)
g <- rep(c("a","b"),10)

ac(x,y)
sil(x,y,g)

[Package gclus version 1.3.1 Index]