body {gclus}	R Documentation

Exploring Relationships in Body Dimensions

Description

This dataset contains 21 body dimension measurements as well as age, weight, height, and gender on 507 individuals. The 247 men and 260 women were primarily individuals in their twenties and thirties, with a scattering of older men and women, all exercising several hours a week.

Measurements were initially taken by Grete Heinz and Louis J. Peterson - at San Jose State University and at the U.S. Naval Postgraduate School in Monterey, California. Later, measurements were taken at dozens of California health and fitness clubs by technicians under the supervision of one of these authors.

Usage

data(body)

Format

This data frame contains the following columns:

Biacrom:

Biacromial diameter (cm)

Biiliac:

Biiliac diameter, or "pelvic breadth" (cm)

Bitro:

Bitrochanteric diameter (cm)

ChestDp:

Chest depth between spine and sternum at nipple level, mid-expiration (cm)

ChestD:

Chest diameter at nipple level, mid-expiration (cm)

ElbowD:

Elbow diameter, sum of two elbows (cm)

WristD:

Wrist diameter, sum of two wrists (cm)

KneeD:

Knee diameter, sum of two knees (cm)

AnkleD:

Ankle diameter, sum of two ankles (cm)

ShoulderG:

Shoulder girth over deltoid muscles (cm)

ChestG:

Chest girth, nipple line in males and just above breast tissue in females, mid-expiration (cm)

WaistG:

Waist girth, narrowest part of torso below the rib cage, average of contracted and relaxed position (cm)

AbdG:

Navel (or "Abdominal") girth at umbilicus and iliac crest, iliac crest as a landmark (cm)

HipG:

Hip girth at level of bitrochanteric diameter (cm)

ThighG:

Thigh girth below gluteal fold, average of right and left girths (cm)

BicepG:

Bicep girth, flexed, average of right and left girths (cm)

ForearmG:

Forearm girth, extended, palm up, average of right and left girths (cm)

KneeG:

Knee girth over patella, slightly flexed position, average of right and left girths (cm)

CalfG:

Calf maximum girth, average of right and left girths (cm)

AnkleG:

Ankle minimum girth, average of right and left girths (cm)

WristG:

Wrist minimum girth, average of right and left girths (cm)

Age:

in years

Weight:

in kg

Height:

in cm

Gender:

1 - male, 0 - female

Source

Heinz, G., Peterson, L.J., Johnson, R.W. and Kerk, C.J. (2003), “Exploring Relationships in Body Dimensions”, Journal of Statistics Education , 11.

References

The data file is taken from http://www.amstat.org/publications/jse/datasets/body.dat This information file is based on http://www.amstat.org/publications/jse/datasets/body.txt

[Package gclus version 1.3.1 Index]