diameter {gclus}	R Documentation

Cluster heterogeneity of 2-d data

Description

Computes measures of cluster heterogeneity of 2-d data, where x and y give the object coordinates.

Usage

diameter(x, y, ...)
star(x, y, ...)
km2(x,y)
gtot(x,y, ...)
gave(x,y, ...)

Arguments

x
is a numeric vector.

y
is a numeric vector.

...
are passed to dist.

Details

diameter computes the cluster diameter- the maximum distance between objects.

star computes the cluster star distance- the smallest total distance from one object to another.

km2 computes the kmeans distance.

gtot computes the sum of all inter-object distances.

gave computes the per-object average of all inter-object distances.

Value

The cluster measure is returned.

Author(s)

Catherine B. Hurley

References

See Gordon, A. D. (1999).“Classification”. Second Edition. London: Chapman and Hall / CRC

See Also

colpairs, cpairs, order.single

Examples

x <- runif(20)
y <- runif(20)
diameter(x,y)

[Package gclus version 1.3.1 Index]