ozone {gclus}	R Documentation

Ozone data from Breiman and Friedman, 1985

Description

This is the Ozone data discussed in Breiman and Friedman (JASA, 1985, p. 580). These data are for 330 days in 1976. All measurements are in the area of Upland, CA, east of Los Angeles.

Usage

data(ozone)

Format

This data frame contains the following columns:

Ozone:

Ozone conc., ppm, at Sandbug AFB.

Temp:

Temperature F. (max?).

InvHt:

Inversion base height, feet

Pres:

Daggett pressure gradient (mm Hg)

Vis:

Visibility (miles)

Hgt:

Vandenburg 500 millibar height (m)

Hum:

Humidity, percent

InvTmp:

Inversion base temperature, degrees F.

Wind:

Wind speed, mph

Source

Breiman, L and Friedman, J. (1985), “Estimating Optimal Transformations for Multiple Regression and Correlation”, Journal of the American Statistical Association, 80, 580-598.

[Package gclus version 1.3.1 Index]