pclen {gclus}	R Documentation

Profile smoothness measures

Description

Computes measures of profile smoothness of 2-d data, where x and y give the object coordinates.

Usage

pclen(x, y)
pcglen(x, y)

Arguments

x
is a numeric vector.

y
is a numeric vector.

Details

pclen computes the total line length in a parallel coordinate plot of x and y.

pcglen computes the average (per object) line length in a parallel coordinate plot where all pairs of objects are connected.

Usually, the data is standardized prior to using these functions.

Value

The panel measure is returned.

Author(s)

Catherine B. Hurley

References

Hurley, Catherine B. “Clustering Visualisations of Multidimensional Data”, Journal of Computational and Graphical Statistics, vol. 13, (4), pp 788-806, 2004.

See Also

cparcoord, colpairs, order.endlink.

Examples

x <- runif(20)
y <- runif(20)
pclen(x,y)

data(state)
mins <- apply(state.x77,2,min)
ranges <- apply(state.x77,2,max) - mins
state.m <- -colpairs(scale(state.x77,mins,ranges), pclen)
state.col <- dmat.color(state.m)
cparcoord(state.x77, panel.color= state.col)
# Get rid of the panels with long line segments (yellow) by reordering:
cparcoord(state.x77, order.endlink(state.m), state.col)

[Package gclus version 1.3.1 Index]