wine {gclus}	R Documentation

Wine recognition data

Description

Data from the machine learning repository. A chemical analysis of 178 Italian wines from three different cultivars yielded 13 measurements. This dataset is often used to test and compare the performance of various classification algorithms.

Usage

data(wine)

Format

This data frame contains the following columns:

Class:

There are 3 classes

Alcohol:

Alcohol

Malic:

Malic acid

Ash:

Ash

Alcalinity:

Alcalinity of ash

Magnesium:

Magnesium

Phenols:

Total phenols

Flavanoids:

Flavanoids

Nonflavanoid:

Nonflavanoid phenols

Proanthocyanins:

Proanthocyanins

Intensity:

Color intensity

Hue:

Hue

OD280:

OD280/OD315 of diluted wines

Proline:

Proline

Source

Forina, M. et al, PARVUS - An Extendible Package for Data Exploration, Classification and Correlation. Institute of Pharmaceutical and Food Analysis and Technologies, Via Brigata Salerno, 16147 Genoa, Italy.

References

Blake, C.L. and Merz, C.J. (1998), UCI Repository of machine learning databases, \ http://www.ics.uci.edu/~mlearn/MLRepository.html. Irvine, CA: University of California, Department of Information and Computer Science.

The database does not list the variable names. These were located at http://www.radwin.org/michael/projects/learning/about-wine.html.

[Package gclus version 1.3.1 Index]